use static_cast to address implicit conversion  warnings JIRA SB-3592
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@100101 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ include/algo/blast/api/cdd_pssm_input.hpp
@@ -141,7 +141,7 @@ public:
     /// Get query length
     /// @return Query length
     ///
-    unsigned int  GetQueryLength(void) {return m_QueryData.size();}
+    unsigned int  GetQueryLength(void) {return static_cast<unsigned int>(m_QueryData.size());}
 
     /// Get scoring matrix name
     /// @return Scoring matrix name
